requirements.txt
numpy==1.21.4
matplotlib==3.5.0
argparse==1.1
sklearn==0.0
pandas==1.4.3
pillow==9.1.0
transformers==4.20.1
torch==1.10.0+cu113
torchvision==0.11.1+cu113
